+++ NEWS.md
@@ -1,3 +1,6 @@
+# witchtools 0.4.5 [devel]
+
+
 # witchtools 0.4.4
 - Write yaml for native regions and mappings for the IIASA database
 - Add r10 region aggregation, based on limits10